requirements.txt
alembic==0.7.5.post2
Flask==0.10.1
Flask-Migrate==1.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
oauthlib==0.7.2
psycopg2==2.6
pytz==2015.2
requests==2.6.2
requests-oauthlib==0.4.2
six==1.9.0
SQLAlchemy==1.0.2
tweepy==3.3.0
Werkzeug==0.10.4
